+++ src/main/java/com/nagarro/multipledbpoc/repository/subscribevehiclemanagement/CategoryMasterRepository.java
@@ -6,10 +6,12 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.nagarro.multipledbpoc.domain.category.CategoryMaster;
+import com.nagarro.multipledbpoc.exception.CategoryIdNotFoundException;
+
 @Repository
 public interface CategoryMasterRepository extends JpaRepository<CategoryMaster, Long> {
-	
-	
+
 	@Query(value = "Select cm.id from public.category_master cm Where  cm.name=:name", nativeQuery = true)
-	Integer findIdByName(@Param("name") String name);
+	Long findIdByName(@Param("name") String name) ;
+//			throws CategoryIdNotFoundException;
 }
